requirements.txt
python-telegram-bot
yandexcloud
boto3
